% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aicsums.R
\name{plot.aicsums}
\alias{plot.aicsums}
\title{Plot sum-of-AIC results}
\usage{
\method{plot}{aicsums}(x, ...)
}
\arguments{
\item{x}{an object of class \code{aicsums}, as produced by \code{\link[optimus]{find_optimal}}.}

\item{...}{additional arguments to pass to \code{\link[graphics]{plot}}.}
}
\description{
S3 \code{\link{plot}} method for sum-of-AIC reuslts from \code{\link[optimus]{find_optimal}}.
}
\examples{

## Example of a find_optimal() call and subsequent plot, and how to interpret
}

