% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aicsums.R
\name{points.aicsums}
\alias{points.aicsums}
\title{Plot more sum-of-AIC results}
\usage{
\method{points}{aicsums}(x, col = sample(1:20, 1), pch = sample(c(1:15,
  17:20), 1), ...)
}
\arguments{
\item{x}{an object of class \code{aicsums}, as produced by \code{\link[optimus]{find_optimal}}.}

\item{col}{point colour - random if not specified}

\item{pch}{point type - random if not specified}

\item{...}{additional arguments to pass to \code{\link[graphics]{points}}.}
}
\value{
Points drawn on the current plot
}
\description{
S3 \code{\link{points}} method for sum-of-AIC reuslts from \code{\link[optimus]{find_optimal}}. Implemented to compare multiple outputs from \code{\link[optimus]{find_optimal}}.
}
\examples{

## see ?find_optimal()
}
\author{
Mitchell Lyons
}
