% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swamps}
\alias{swamps}
\title{Dharawal National Park Upland Heath Swamps Plot Network}
\format{A data frame with 54 rows and 171 variables:
\describe{
  \item{transect}{transect identifyer, that is, transect site (0.5 x 0.5 m quadrats)}
  \item{remaining columns}{frequency counts (out of 30) for all vascular plant taxa; taxonomy in metadata link below}
  ...
}}
\source{
\url{http://www.ltern.org.au/knb/metacat?action=read&qformat=html&docid=ltern.84.15}
}
\usage{
swamps
}
\description{
The Upland Heath Swamps Plot Network Vegetation Structure and Floristics Data Package contains information on the vegetation structure and species present in 60 established swamp monitoring sites in upland swamps scattered throughout the study area (Keith and Myerscough 1993). Each site is sampled in nine combinations of moisture-by-vegetation structure strata. The Upland Heath Swamps Plot Network research plots commenced in 1983 and have been revisited in 2004, 2009 and again in 2014. More detail can be found at \url{http://www.ltern.org.au/index.php/ltern-plot-networks/upland-health-swamps}.
}
\keyword{datasets}

