% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aicsums.R
\name{plot.aicsums}
\alias{plot.aicsums}
\title{Plot sum-of-AIC results}
\usage{
\method{plot}{aicsums}(x, col = "black", pch = 16, ...)
}
\arguments{
\item{x}{an object of class \code{aicsums}, as produced by \code{\link[optimus]{find_optimal}}.}

\item{col}{point colour}

\item{pch}{point type}

\item{...}{additional arguments to pass to \code{\link[graphics]{plot}}.}
}
\value{
A plot is drawn on the current graphics device
}
\description{
S3 \code{\link{plot}} method for sum-of-AIC results from \code{\link[optimus]{find_optimal}}.
}
\examples{

## see ?find_optimal()
}
\author{
Mitchell Lyons
}
